Retrieve all parameters from the form.
Otherwise, when you refresh the form and it looks filled in,
you'll get an error.  No change events have fired, so no values
are populated in the MeetingTicker object.  That's just not
friendly.

Also, I've changed the action attributes of the forms to '#'.  javascript:false
doesn't work, and if the event fires it looks really ugly.